requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.27.0
requests>=2.31.0
resend>=0.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
